requirements.txt
azure-cognitiveservices-speech==1.38.0
pyqt5==5.15.10
openai==1.35.7
python-dotenv==1.0.1